backend/functions/embedding-service/requirements.txt
functions-framework==3.*
openai==1.3.0
google-cloud-firestore==2.13.1
sentence-transformers==2.2.2
